@@ -24,7 +24,7 @@
   "scripts": {
     "build": "rc-tools run build",
     "gh-pages": "rc-tools run gh-pages",
-    "start": "rc-server",
+    "start": "rc-tools run server",
     "pub": "rc-tools run pub",
     "lint": "rc-tools run lint",
     "karma": "rc-tools run karma",
@@ -35,16 +35,17 @@
   "devDependencies": {
     "expect.js": "~0.3.1",
     "pre-commit": "1.x",
-    "rc-server": "3.x",
-    "rc-tools": "4.x",
-    "react": "^0.14.0",
-    "react-addons-test-utils": "~0.14.0",
-    "react-dom": "~0.14.0"
+    "rc-tools": "5.x",
+    "react": "15.x",
+    "react-addons-test-utils": "15.x",
+    "react-dom": "~15.x",
+    "xtemplate": "4.6.0"
   },
   "pre-commit": [
     "lint"
   ],
   "dependencies": {
-    "classnames": "^2.2.1"
+    "classnames": "^2.2.1",
+    "prop-types": "^15.5.6"
   }
 }
