@@ -1,48 +1,32 @@
-const React = require('react');
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 const classNames = require('classnames');
 
 function noop() {
 }
 
-const Switch = React.createClass({
-  propTypes: {
-    className: React.PropTypes.string,
-    prefixCls: React.PropTypes.string,
-    disabled: React.PropTypes.bool,
-    checkedChildren: React.PropTypes.any,
-    unCheckedChildren: React.PropTypes.any,
-    onChange: React.PropTypes.func,
-    onMouseUp: React.PropTypes.func,
-  },
-  getDefaultProps() {
-    return {
-      prefixCls: 'rc-switch',
-      checkedChildren: null,
-      unCheckedChildren: null,
-      className: '',
-      defaultChecked: false,
-      onChange: noop,
-    };
-  },
-  getInitialState() {
-    const props = this.props;
+class Switch extends Component {
+
+  constructor(props) {
+    super(props);
+
     let checked = false;
     if ('checked' in props) {
       checked = !!props.checked;
     } else {
       checked = !!props.defaultChecked;
     }
-    return {
-      checked,
-    };
-  },
+    this.state = { checked };
+  }
+
   componentWillReceiveProps(nextProps) {
     if ('checked' in nextProps) {
       this.setState({
         checked: !!nextProps.checked,
       });
     }
-  },
+  }
+
   setChecked(checked) {
     if (!('checked' in this.props)) {
       this.setState({
@@ -50,11 +34,13 @@ const Switch = React.createClass({
       });
     }
     this.props.onChange(checked);
-  },
+  }
+
   toggle() {
     const checked = !this.state.checked;
     this.setChecked(checked);
-  },
+  }
+
   handleKeyDown(e) {
     if (e.keyCode === 37) {
       this.setChecked(false);
@@ -62,7 +48,8 @@ const Switch = React.createClass({
     if (e.keyCode === 39) {
       this.setChecked(true);
     }
-  },
+  }
+
   // Handle auto focus when click switch in Chrome
   handleMouseUp(e) {
     if (this.refs.node) {
@@ -71,9 +58,10 @@ const Switch = React.createClass({
     if (this.props.onMouseUp) {
       this.props.onMouseUp(e);
     }
-  },
+  }
+
   render() {
-    const {className, prefixCls, disabled,
+    const { className, prefixCls, disabled,
       checkedChildren, unCheckedChildren, ...restProps } = this.props;
     const checked = this.state.checked;
     const switchClassName = classNames({
@@ -87,15 +75,34 @@ const Switch = React.createClass({
         className={switchClassName}
         tabIndex={disabled ? -1 : 0}
         ref="node"
-        onKeyDown={this.handleKeyDown}
-        onClick={disabled ? noop : this.toggle}
-        onMouseUp={this.handleMouseUp}>
+        onKeyDown={this.handleKeyDown.bind(this)}
+        onClick={disabled ? noop : this.toggle.bind(this)}
+        onMouseUp={ this.handleMouseUp.bind(this) }>
         <span className={`${prefixCls}-inner`}>
           {checked ? checkedChildren : unCheckedChildren}
         </span>
       </span>
     );
-  },
-});
+  }
+}
 
-module.exports = Switch;
+Switch.propTypes = {
+  className: PropTypes.string,
+  prefixCls: PropTypes.string,
+  disabled: PropTypes.bool,
+  checkedChildren: PropTypes.any,
+  unCheckedChildren: PropTypes.any,
+  onChange: PropTypes.func,
+  onMouseUp: PropTypes.func,
+};
+
+Switch.defaultProps = {
+  prefixCls: 'rc-switch',
+  checkedChildren: null,
+  unCheckedChildren: null,
+  className: '',
+  defaultChecked: false,
+  onChange: noop,
+};
+
+export default Switch;
